Refactor & Add IndicatorDirection component
