+++ client/src/components/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+export { default as IndicatorDirection } from './IndicatorDirection';
