+++ client/src/pages/Mockup.vue
@@ -47,19 +47,7 @@
       </div>
 
       <div class="bloque-motor tercio">
-        <div class="direction">
-          direction<br />
-          <div class="bar">
-            <div class="side left">
-              <div class="fill" style="width: 57%;"></div>
-            </div><!--
-            --><div class="middle"></div><!--
-            --><div class="side right">
-              <div class="fill" style="width: 2%;"></div>
-            </div>
-          </div>
-        </div>
-
+        <indicator-direction />
         <div class="engine">
           <div class="engine-half engine-oil-tem">190</div>
           <div class="engine-half engine-water-tem">85</div>
@@ -102,8 +90,13 @@
 </template>
 
 <script>
+import { IndicatorDirection } from '../components/index';
+
 export default {
   name: 'Mockup',
+  components: {
+    IndicatorDirection,
+  },
   data() {
     return {
       msg: 'Welcome to Your Vue.js App',
@@ -274,39 +267,6 @@ export default {
   .bloque-motor {
     z-index: 2;
   }
-
-  .direction {
-    height: 10%;
-    font-size: 1.5em;
-    padding: 2vh 0;
-    width: 100%;
-  }
-    .direction .bar {
-      background-color: lightgrey;
-      height: 1vh;
-      width: 60%;
-    }
-      .direction .bar .side {
-        height: 100%;
-        width: calc(48% - 1vh);
-        margin-top: -2vh;
-      }
-        .direction .bar .side.left {
-          text-align: right;
-        }
-        .direction .bar .side.right {
-          text-align: left;
-        }
-        .direction .bar .side .fill {
-          background-color: green;
-          height: 100%;
-          margin-top: -2vh;
-        }
-      .direction .bar .middle {
-        background-color: red;
-        height: 100%;
-        width: 1vh;
-      }
 
   .engine {
     height: 30%;
